capstone/requirements.txt
streamlit
opencv-python-headless
numpy
scikit-image
Pillow
fpdf

